+++ lib/HiPay/Fullservice/Gateway/Model/SecuritySettings.php
@@ -13,13 +13,14 @@
  * @license        http://www.apache.org/licenses/LICENSE-2.0 Apache 2.0 Licence
  *
  */
+
 namespace HiPay\Fullservice\Gateway\Model;
 
 use HiPay\Fullservice\Model\AbstractModel;
 
 /**
  * Security Settings Model
- * 
+ *
  * @package HiPay\Fullservice
  * @author Etienne Landais <elandais@hipay.com>
  * @copyright Copyright (c) 2018 - HiPay
@@ -29,34 +30,33 @@ use HiPay\Fullservice\Model\AbstractModel;
  */
 class SecuritySettings extends AbstractModel
 {
-	/**
-	 * @var string hash algorithm used for signature
-	 */
-	private $hashingAlgorithm;
+    /**
+     * @var string hash algorithm used for signature
+     */
+    protected $hashingAlgorithm;
 
-	/**
-	 * SecuritySettings constructor.
-	 * @param string $hashingAlgorithm
-	 */
-	public function __construct($hashingAlgorithm)
-	{
-		$this->hashingAlgorithm = $hashingAlgorithm;
-	}
+    /**
+     * SecuritySettings constructor.
+     * @param string $hashingAlgorithm
+     */
+    public function __construct($hashingAlgorithm)
+    {
+        $this->hashingAlgorithm = $hashingAlgorithm;
+    }
 
-	/**
-	 * @return string
-	 */
-	public function getHashingAlgorithm()
-	{
-		return $this->hashingAlgorithm;
-	}
+    /**
+     * @return string
+     */
+    public function getHashingAlgorithm()
+    {
+        return $this->hashingAlgorithm;
+    }
 
-	/**
-	 * @param string $hashingAlgorithm
-	 */
-	public function setHashingAlgorithm($hashingAlgorithm)
-	{
-		$this->hashingAlgorithm = $hashingAlgorithm;
-	}
-
+    /**
+     * @param string $hashingAlgorithm
+     */
+    public function setHashingAlgorithm($hashingAlgorithm)
+    {
+        $this->hashingAlgorithm = $hashingAlgorithm;
+    }
 }