+++ lib/HiPay/Fullservice/Request/AbstractRequest.php
@@ -13,19 +13,21 @@
  * @license        http://www.apache.org/licenses/LICENSE-2.0 Apache 2.0 Licence
  *
  */
+
 namespace HiPay\Fullservice\Request;
 
 /**
  * Model request abstract.
- * All request object sended by Gateway or Secure Vault client must implement it
+ * All request object sended by Gateway client must implement it
  *
  * @package     HiPay\Fullservice
- * @author 		Kassim Belghait <kassim@sirateck.com>
+ * @author        Kassim Belghait <kassim@sirateck.com>
  * @copyright   Copyright (c) 2016 - HiPay
  * @license     http://www.apache.org/licenses/LICENSE-2.0 Apache 2.0 License
- * @link 		https://github.com/hipay/hipay-fullservice-sdk-php
+ * @link        https://github.com/hipay/hipay-fullservice-sdk-php
  */
-abstract class AbstractRequest implements RequestInterface{
+abstract class AbstractRequest implements RequestInterface
+{
 
-		
+
 }